requirements.txt
pytest
requests
allure-pytest
playwright